Refactored files to team.asd
