+++ src/main/java/team/asd/JavaInternApplication.java
@@ -1,4 +1,4 @@
-package javaintern;
+package team.asd;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
