+++ src/main/java/team/asd/constant/ReservationState.java
@@ -1,4 +1,4 @@
-package javaintern.constant;
+package team.asd.constant;
 
 public enum ReservationState {
 	Initial, Confirmed, FullyPaid, Cancelled
